Move minimal playlist entries reducer to playlist.live.entries
